Use custom SVG creation logic
The initial way of creating the SVG file was using
an external library in which it was not clear on how
certain calculations were done and thus the resulting
scales had some issues with the padding and sometimes
with the cutting rectangle.

This commit now switches the SVG generation to SkiaSharp
for the font paths and custom XML handling to create the
SVG file. All calculations can now be seen in code
(except for creating paths from font files).

Also some tests were added to avoid future regressions.